BigInt fixes (#5726)

+++ src/lib/settings/prisma.ts
@@ -59,7 +59,7 @@ export function convertStoredActivityToFlatActivity(activity: Activity): Activit
  * ⚠️ Uses queryRawUnsafe
  */
 export async function countUsersWithItemInCl(itemID: number, ironmenOnly: boolean) {
-	const query = `SELECT COUNT(id)
+	const query = `SELECT COUNT(id)::int
 				   FROM users
 				   WHERE ("collectionLogBank"->>'${itemID}') IS NOT NULL 
 				   AND ("collectionLogBank"->>'${itemID}')::int >= 1
